Still refactoring

+++ src/benchmark/bench_sample_sort.rs
@@ -15,11 +15,11 @@ pub fn benchmark_sample_sort(config: &mut Config) {
     let args: Vec<String> = env::args().collect();
     if args.len() <= 1 { panic!("not enough arguments"); }
 
-    let rankn: usize = config.rankn as usize;
-    let rank: usize = config.rank as usize;
+    let rankn: u32 = config.rankn as u32;
+    let rank: u32 = config.rank as u32;
 
-    let n: i32 = args[1].clone().parse().unwrap();
-    let size: i32 = n * rankn;
+    let n: u32 = args[1].clone().parse().unwrap();
+    let size: u32 = n * rankn;
 
     let mut data: Vec<i32> = Vec::new();
     let mut rng: StdRng = SeedableRng::from_seed([rankn as u8; 32]);
@@ -31,7 +31,7 @@ pub fn benchmark_sample_sort(config: &mut Config) {
 
     // scattering data
     let mut local: Vec<GlobalPointer<i32>> = Vec::new();
-    local.resize(rankn, GlobalPointer::null());
+    local.resize(rankn as usize, GlobalPointer::null());
     local[rank] = GlobalPointer::init(config, n as usize);
     for i in 0 .. n {
         (local[rank] + i as isize).rput(data[rank * n + i]);
